--- app/src/main/java/huiwoo/jo/mvvm/project/presentation/detail/DetailNewsViewModel.kt
@@ -1,4 +1,0 @@
-package huiwoo.jo.mvvm.project.presentation.detail
-
-class DetailNewsViewModel {
-}
