--- app/src/main/java/huiwoo/jo/mvvm/project/presentation/main/MainNewsViewModel.kt
@@ -1,4 +1,0 @@
-package huiwoo.jo.mvvm.project.presentation.main
-
-class MainNewsViewModel {
-}
