+++ app/src/main/java/huiwoo/jo/mvvm/project/presentation/news/DetailFragment.kt
@@ -1,4 +1,4 @@
-package huiwoo.jo.mvvm.project.presentation.main
+package huiwoo.jo.mvvm.project.presentation.news
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -14,10 +14,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [MainNewsFragment.newInstance] factory method to
+ * Use the [DetailFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class MainNewsFragment : Fragment() {
+class DetailFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -35,7 +35,7 @@ class MainNewsFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_main_news, container, false)
+        return inflater.inflate(R.layout.fragment_detail_news, container, false)
     }
 
     companion object {
@@ -45,13 +45,12 @@ class MainNewsFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment MainNewsFragment.
+         * @return A new instance of fragment DetailFragment.
          */
         // TODO: Rename and change types and number of parameters
-        fun newInstance() = MainNewsFragment()
-        const val TAG = "MainNewsFragment"
+        @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            MainNewsFragment().apply {
+            DetailFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
