+++ app/src/main/java/huiwoo/jo/mvvm/project/presentation/news/DetailViewModel.kt
@@ -1,0 +1,4 @@
+package huiwoo.jo.mvvm.project.presentation.news
+
+class DetailViewModel {
+}
