+++ app/src/main/res/layout/fragment_detail_news.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.detail.DetailNewsFragment">
+    tools:context=".presentation.news.DetailFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
